requirements.txt
fastapi
uvicorn
sqlalchemy
python-multipart
passlib
bcrypt
tensorflow
numpy
opencv-python
pillow